+++ dev/index.js
@@ -4,14 +4,17 @@ const loop = require('raf-loop')
 //const shader = require('./../index.js').shaderGenerator
 
 function init () {
+//  exampleNonGlobal()
+
+//  exampleExtendTransforms()
   var hydra = new Hydra()
-  window.hydra = hydra
-
-  exampleAddFunction(hydra)
-//  exampleScreen()
-
-exampleCustomCanvas()
-exampleSmoothing()
+//   window.hydra = hydra
+//
+//   exampleAddFunction(hydra)
+// //  exampleScreen()
+//
+// exampleCustomCanvas()
+// exampleSmoothing()
   ///var generator = new shader()
 
 
@@ -57,6 +60,45 @@ function exampleEasingFunctions() {
   //  // shape([4, 3, 2].smooth()).out()
 }
 
+function exampleNonGlobal() {
+    var hydra = new Hydra({ makeGlobal: false })
+    hydra.synth.fps = 10
+    hydra.synth.osc().out()
+    window.hydra = hydra
+}
+
+function exampleExtendTransforms() {
+  var hydra = new Hydra({
+    extendTransforms: {
+      name: 'myOsc', // name that will be used to access function as well as within glsl
+      type: 'src', // can be src: vec4(vec2 _st), coord: vec2(vec2 _st), color: vec4(vec4 _c0), combine: vec4(vec4 _c0, vec4 _c1), combineCoord: vec2(vec2 _st, vec4 _c0)
+      inputs: [
+        {
+          name: 'freq',
+          type: 'float', // 'float'   //, 'texture', 'vec4'
+          default: 0.2
+        },
+        {
+              name: 'sync',
+              type: 'float',
+              default: 0.1
+            },
+            {
+              name: 'offset',
+              type: 'float',
+              default: 0.0
+            }
+      ],    glsl: `
+         vec2 st = _st;
+        float r = sin((st.x-offset*20./freq-time*sync)*freq)*0.5  + 0.5;
+        float g = sin((st.x+time*sync)*freq)*0.5 + 0.5;
+        float b = sin((st.x+offset/freq+time*sync)*freq)*0.5  + 0.5;
+        return vec4(r, g, b, 1.0);
+       `}
+    })
+    myOsc(10, 0.2, 0.8).out()
+
+}
 
 function exampleAddFunction(hydra) {
  // example custom function
@@ -135,6 +177,10 @@ function exampleSmoothing() {
    .add(shapes[1])
    .add(shapes[2])
    .out(o0)
+}
+
+function exampleSetResolution() {
+  setResolution(20, 20)
 }
 
 window.onload = init