+++ src/generator-factory.js
@@ -1,11 +1,11 @@
-const glslTransforms = require('./glsl/composable-glsl-functions.js')
+const glslTransforms = require('./glsl/glsl-functions.js')
 const GlslSource = require('./glsl-source.js')
 
 class GeneratorFactory {
   constructor ({
       defaultUniforms,
       defaultOutput,
-      extendTransforms = (x => x),
+      extendTransforms = [],
       changeListener = (() => {})
     } = {}
     ) {
@@ -28,39 +28,41 @@ class GeneratorFactory {
       }
     })()
 
-    let functions = {}
-    const addTransforms = (transforms) =>
-      Object.entries(transforms).forEach(([method, transform]) => {
-        functions[method] = transform
-      })
+    let functions = glslTransforms
+    // const addTransforms = (transforms) =>
+    //   Object.entries(transforms).forEach(([method, transform]) => {
+    //     functions[method] = transform
+    //   })
 
-    addTransforms(glslTransforms)
-    //addTransforms(renderpassFunctions)
-
-    if (typeof this.extendTransforms === 'function') {
-      functions = this.extendTransforms(functions)
-    } else if (Array.isArray(this.extendTransforms)) {
-      addTransforms(this.extendTransforms.reduce((h, transform) => {
-        h[transform.name] = transform
-        return h
-      }, {}))
+    // const addTransforms =
+    //
+    // addTransforms(glslTransforms)
+    // //addTransforms(renderpassFunctions)
+    //
+    // if (typeof this.extendTransforms === 'function') {
+    //   functions = this.extendTransforms(functions)
+    // } else
+    if (Array.isArray(this.extendTransforms)) {
+      functions.concat(this.extendTransforms)
     } else if (typeof this.extendTransforms === 'object') {
-      addTransforms(this.extendTransforms)
+      functions.push(this.extendTransforms)
     }
 
-    Object.entries(functions).forEach(([method, transform]) => {
-      if (typeof transform.glsl_return_type === 'undefined' && transform.glsl) {
-        transform.glsl_return_type = transform.glsl.replace(new RegExp(`^(?:[\\s\\S]*\\W)?(\\S+)\\s+${method}\\s*[(][\\s\\S]*`, 'ugm'), '$1')
-      }
+  //  console.log(functions)
+    //
+    // Object.entries(functions).forEach(([method, transform]) => {
+    //   if (typeof transform.glsl_return_type === 'undefined' && transform.glsl) {
+    //     transform.glsl_return_type = transform.glsl.replace(new RegExp(`^(?:[\\s\\S]*\\W)?(\\S+)\\s+${method}\\s*[(][\\s\\S]*`, 'ugm'), '$1')
+    //   }
+    //
+    //   functions[method] = this._addMethod(method, transform)
+    // })
 
-      functions[method] = this._addMethod(method, transform)
-    })
-
-    return functions
+    return functions.map((transform) => this.setFunction(transform))
  }
 
  _addMethod (method, transform) {
-   console.log('adding', method, transform)
+//   console.log('adding', method, transform)
     this.glslTransforms[method] = transform
     if (transform.type === 'src') {
       const func = (...args) => new this.sourceClass({
@@ -85,8 +87,8 @@ class GeneratorFactory {
 
   setFunction(obj) {
     var processedGlsl = processGlsl(obj)
-    console.log(processedGlsl)
-    this._addMethod(obj.name, processedGlsl)
+  //  console.log(processedGlsl)
+    if(processedGlsl) this._addMethod(obj.name, processedGlsl)
   }
 }
 
@@ -109,9 +111,9 @@ const typeLookup = {
     returnType: 'vec4',
     args: ['vec4 _c0', 'vec4 _c1']
   },
-  'combineCoords': {
+  'combineCoord': {
     returnType: 'vec2',
-    args: ['vec2 _st', 'vec4 c0']
+    args: ['vec2 _st', 'vec4 _c0']
   }
 }
 // expects glsl of format
@@ -155,22 +157,33 @@ const typeLookup = {
 
 function processGlsl(obj) {
   let t = typeLookup[obj.type]
+  if(t) {
   let baseArgs = t.args.map((arg) => arg).join(", ")
   // @todo: make sure this works for all input types, add validation
   let customArgs = obj.inputs.map((input) => `${input.type} ${input.name}`).join(', ')
   let args = `${baseArgs}${customArgs.length > 0 ? ', '+ customArgs: ''}`
-  console.log('args are ', args)
-  if(t) {
+//  console.log('args are ', args)
+
     let glslFunction =
 `
   ${t.returnType} ${obj.name}(${args}) {
       ${obj.glsl}
   }
 `
-    console.log('function', glslFunction)
+  //  console.log('function', glslFunction)
+    // {
+    //   name: 'color',
+    //   type: 'vec4'
+    // },
+
+  // add extra input to beginning for backward combatibility @todo update compiler so this is no longer necessary
+    if(obj.type === 'combine' || obj.type === 'combineCoord') obj.inputs.unshift({
+        name: 'color',
+        type: 'vec4'
+      })
     return Object.assign({}, obj, { glsl: glslFunction})
   } else {
-    console.error(`type ${obj.type} not recognized`)
+    console.warn(`type ${obj.type} not recognized`)
   }
 
 }